+++ tgui/packages/tgui/styles/components/Dimmer.scss
@@ -23,6 +23,6 @@ $background-dimness: 0.75 !default;
   // Darken the inner section to be more readable on colourful backgrounds.
   &__inner {
     background-color: rgba(0, 0, 0, $background-dimness);
-    border: 2em solid rgba(0, 0, 0, $background-dimness);
+    border: 2em solid rgba(0, 0, 0, 0);
   }
 }
